Repot multisig to make utility stateless.
