+++ bin/node/runtime/Cargo.toml
@@ -59,6 +59,7 @@ pallet-im-online = { version = "2.0.0-rc2", default-features = false, path = "..
 pallet-indices = { version = "2.0.0-rc2", default-features = false, path = "../../../frame/indices" }
 pallet-identity = { version = "2.0.0-rc2", default-features = false, path = "../../../frame/identity" }
 pallet-membership = { version = "2.0.0-rc2", default-features = false, path = "../../../frame/membership" }
+pallet-multisig = { version = "2.0.0-rc2", default-features = false, path = "../../../frame/multisig" }
 pallet-offences = { version = "2.0.0-rc2", default-features = false, path = "../../../frame/offences" }
 pallet-offences-benchmarking = { version = "2.0.0-rc2", path = "../../../frame/offences/benchmarking", default-features = false, optional = true }
 pallet-proxy = { version = "2.0.0-rc2", default-features = false, path = "../../../frame/proxy" }
@@ -108,6 +109,7 @@ std = [
 	"pallet-indices/std",
 	"sp-inherents/std",
 	"pallet-membership/std",
+	"pallet-multisig/std",
 	"pallet-identity/std",
 	"node-primitives/std",
 	"sp-offchain/std",
@@ -151,6 +153,7 @@ runtime-benchmarks = [
 	"pallet-elections-phragmen/runtime-benchmarks",
 	"pallet-identity/runtime-benchmarks",
 	"pallet-im-online/runtime-benchmarks",
+	"pallet-multisig/runtime-benchmarks",
 	"pallet-proxy/runtime-benchmarks",
 	"pallet-scheduler/runtime-benchmarks",
 	"pallet-society/runtime-benchmarks",
