+++ frame/multisig/Cargo.toml
@@ -1,0 +1,44 @@
+[package]
+name = "pallet-multisig"
+version = "2.0.0-rc2"
+authors = ["Parity Technologies <admin@parity.io>"]
+edition = "2018"
+license = "Apache-2.0"
+homepage = "https://substrate.dev"
+repository = "https://github.com/paritytech/substrate/"
+description = "FRAME multi-signature dispatch pallet"
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[dependencies]
+serde = { version = "1.0.101", optional = true }
+codec = { package = "parity-scale-codec", version = "1.3.0", default-features = false }
+frame-support = { version = "2.0.0-rc2", default-features = false, path = "../support" }
+frame-system = { version = "2.0.0-rc2", default-features = false, path = "../system" }
+sp-core = { version = "2.0.0-rc2", default-features = false, path = "../../primitives/core" }
+sp-runtime = { version = "2.0.0-rc2", default-features = false, path = "../../primitives/runtime" }
+sp-std = { version = "2.0.0-rc2", default-features = false, path = "../../primitives/std" }
+sp-io = { version = "2.0.0-rc2", default-features = false, path = "../../primitives/io" }
+
+frame-benchmarking = { version = "2.0.0-rc2", default-features = false, path = "../benchmarking", optional = true }
+
+[dev-dependencies]
+sp-core = { version = "2.0.0-rc2", path = "../../primitives/core" }
+pallet-balances = { version = "2.0.0-rc2", path = "../balances" }
+
+[features]
+default = ["std"]
+std = [
+	"serde",
+	"codec/std",
+	"sp-runtime/std",
+	"frame-support/std",
+	"frame-system/std",
+	"sp-io/std",
+	"sp-std/std"
+]
+runtime-benchmarks = [
+	"frame-benchmarking",
+	"frame-support/runtime-benchmarks",
+]
