+++ frame/utility/src/tests.rs
@@ -37,7 +37,7 @@ impl_outer_event! {
 	pub enum TestEvent for Test {
 		system<T>,
 		pallet_balances<T>,
-		utility<T>,
+		utility,
 	}
 }
 impl_outer_dispatch! {
@@ -111,10 +111,6 @@ impl Filter<Call> for TestIsCallable {
 impl Trait for Test {
 	type Event = TestEvent;
 	type Call = Call;
-	type Currency = Balances;
-	type MultisigDepositBase = MultisigDepositBase;
-	type MultisigDepositFactor = MultisigDepositFactor;
-	type MaxSignatories = MaxSignatories;
 	type IsCallable = TestIsCallable;
 }
 type System = frame_system::Module<Test>;
@@ -140,264 +136,6 @@ fn last_event() -> TestEvent {
 
 fn expect_event<E: Into<TestEvent>>(e: E) {
 	assert_eq!(last_event(), e.into());
-}
-
-fn now() -> Timepoint<u64> {
-	Utility::timepoint()
-}
-
-#[test]
-fn multisig_deposit_is_taken_and_returned() {
-	new_test_ext().execute_with(|| {
-		let multi = Utility::multi_account_id(&[1, 2, 3][..], 2);
-		assert_ok!(Balances::transfer(Origin::signed(1), multi, 5));
-		assert_ok!(Balances::transfer(Origin::signed(2), multi, 5));
-		assert_ok!(Balances::transfer(Origin::signed(3), multi, 5));
-
-		let call = Box::new(Call::Balances(BalancesCall::transfer(6, 15)));
-		assert_ok!(Utility::as_multi(Origin::signed(1), 2, vec![2, 3], None, call.clone()));
-		assert_eq!(Balances::free_balance(1), 2);
-		assert_eq!(Balances::reserved_balance(1), 3);
-
-		assert_ok!(Utility::as_multi(Origin::signed(2), 2, vec![1, 3], Some(now()), call));
-		assert_eq!(Balances::free_balance(1), 5);
-		assert_eq!(Balances::reserved_balance(1), 0);
-	});
-}
-
-#[test]
-fn cancel_multisig_returns_deposit() {
-	new_test_ext().execute_with(|| {
-		let call = Box::new(Call::Balances(BalancesCall::transfer(6, 15)));
-		let hash = call.using_encoded(blake2_256);
-		assert_ok!(Utility::approve_as_multi(Origin::signed(1), 3, vec![2, 3], None, hash.clone()));
-		assert_ok!(Utility::approve_as_multi(Origin::signed(2), 3, vec![1, 3], Some(now()), hash.clone()));
-		assert_eq!(Balances::free_balance(1), 6);
-		assert_eq!(Balances::reserved_balance(1), 4);
-		assert_ok!(
-			Utility::cancel_as_multi(Origin::signed(1), 3, vec![2, 3], now(), hash.clone()),
-		);
-		assert_eq!(Balances::free_balance(1), 10);
-		assert_eq!(Balances::reserved_balance(1), 0);
-	});
-}
-
-#[test]
-fn timepoint_checking_works() {
-	new_test_ext().execute_with(|| {
-		let multi = Utility::multi_account_id(&[1, 2, 3][..], 2);
-		assert_ok!(Balances::transfer(Origin::signed(1), multi, 5));
-		assert_ok!(Balances::transfer(Origin::signed(2), multi, 5));
-		assert_ok!(Balances::transfer(Origin::signed(3), multi, 5));
-
-		let call = Box::new(Call::Balances(BalancesCall::transfer(6, 15)));
-		let hash = call.using_encoded(blake2_256);
-
-		assert_noop!(
-			Utility::approve_as_multi(Origin::signed(2), 2, vec![1, 3], Some(now()), hash.clone()),
-			Error::<Test>::UnexpectedTimepoint,
-		);
-
-		assert_ok!(Utility::approve_as_multi(Origin::signed(1), 2, vec![2, 3], None, hash));
-
-		assert_noop!(
-			Utility::as_multi(Origin::signed(2), 2, vec![1, 3], None, call.clone()),
-			Error::<Test>::NoTimepoint,
-		);
-		let later = Timepoint { index: 1, .. now() };
-		assert_noop!(
-			Utility::as_multi(Origin::signed(2), 2, vec![1, 3], Some(later), call.clone()),
-			Error::<Test>::WrongTimepoint,
-		);
-	});
-}
-
-#[test]
-fn multisig_2_of_3_works() {
-	new_test_ext().execute_with(|| {
-		let multi = Utility::multi_account_id(&[1, 2, 3][..], 2);
-		assert_ok!(Balances::transfer(Origin::signed(1), multi, 5));
-		assert_ok!(Balances::transfer(Origin::signed(2), multi, 5));
-		assert_ok!(Balances::transfer(Origin::signed(3), multi, 5));
-
-		let call = Box::new(Call::Balances(BalancesCall::transfer(6, 15)));
-		let hash = call.using_encoded(blake2_256);
-		assert_ok!(Utility::approve_as_multi(Origin::signed(1), 2, vec![2, 3], None, hash));
-		assert_eq!(Balances::free_balance(6), 0);
-
-		assert_ok!(Utility::as_multi(Origin::signed(2), 2, vec![1, 3], Some(now()), call));
-		assert_eq!(Balances::free_balance(6), 15);
-	});
-}
-
-#[test]
-fn multisig_3_of_3_works() {
-	new_test_ext().execute_with(|| {
-		let multi = Utility::multi_account_id(&[1, 2, 3][..], 3);
-		assert_ok!(Balances::transfer(Origin::signed(1), multi, 5));
-		assert_ok!(Balances::transfer(Origin::signed(2), multi, 5));
-		assert_ok!(Balances::transfer(Origin::signed(3), multi, 5));
-
-		let call = Box::new(Call::Balances(BalancesCall::transfer(6, 15)));
-		let hash = call.using_encoded(blake2_256);
-		assert_ok!(Utility::approve_as_multi(Origin::signed(1), 3, vec![2, 3], None, hash.clone()));
-		assert_ok!(Utility::approve_as_multi(Origin::signed(2), 3, vec![1, 3], Some(now()), hash.clone()));
-		assert_eq!(Balances::free_balance(6), 0);
-
-		assert_ok!(Utility::as_multi(Origin::signed(3), 3, vec![1, 2], Some(now()), call));
-		assert_eq!(Balances::free_balance(6), 15);
-	});
-}
-
-#[test]
-fn cancel_multisig_works() {
-	new_test_ext().execute_with(|| {
-		let call = Box::new(Call::Balances(BalancesCall::transfer(6, 15)));
-		let hash = call.using_encoded(blake2_256);
-		assert_ok!(Utility::approve_as_multi(Origin::signed(1), 3, vec![2, 3], None, hash.clone()));
-		assert_ok!(Utility::approve_as_multi(Origin::signed(2), 3, vec![1, 3], Some(now()), hash.clone()));
-		assert_noop!(
-			Utility::cancel_as_multi(Origin::signed(2), 3, vec![1, 3], now(), hash.clone()),
-			Error::<Test>::NotOwner,
-		);
-		assert_ok!(
-			Utility::cancel_as_multi(Origin::signed(1), 3, vec![2, 3], now(), hash.clone()),
-		);
-	});
-}
-
-#[test]
-fn multisig_2_of_3_as_multi_works() {
-	new_test_ext().execute_with(|| {
-		let multi = Utility::multi_account_id(&[1, 2, 3][..], 2);
-		assert_ok!(Balances::transfer(Origin::signed(1), multi, 5));
-		assert_ok!(Balances::transfer(Origin::signed(2), multi, 5));
-		assert_ok!(Balances::transfer(Origin::signed(3), multi, 5));
-
-		let call = Box::new(Call::Balances(BalancesCall::transfer(6, 15)));
-		assert_ok!(Utility::as_multi(Origin::signed(1), 2, vec![2, 3], None, call.clone()));
-		assert_eq!(Balances::free_balance(6), 0);
-
-		assert_ok!(Utility::as_multi(Origin::signed(2), 2, vec![1, 3], Some(now()), call));
-		assert_eq!(Balances::free_balance(6), 15);
-	});
-}
-
-#[test]
-fn multisig_2_of_3_as_multi_with_many_calls_works() {
-	new_test_ext().execute_with(|| {
-		let multi = Utility::multi_account_id(&[1, 2, 3][..], 2);
-		assert_ok!(Balances::transfer(Origin::signed(1), multi, 5));
-		assert_ok!(Balances::transfer(Origin::signed(2), multi, 5));
-		assert_ok!(Balances::transfer(Origin::signed(3), multi, 5));
-
-		let call1 = Box::new(Call::Balances(BalancesCall::transfer(6, 10)));
-		let call2 = Box::new(Call::Balances(BalancesCall::transfer(7, 5)));
-
-		assert_ok!(Utility::as_multi(Origin::signed(1), 2, vec![2, 3], None, call1.clone()));
-		assert_ok!(Utility::as_multi(Origin::signed(2), 2, vec![1, 3], None, call2.clone()));
-		assert_ok!(Utility::as_multi(Origin::signed(3), 2, vec![1, 2], Some(now()), call2));
-		assert_ok!(Utility::as_multi(Origin::signed(3), 2, vec![1, 2], Some(now()), call1));
-
-		assert_eq!(Balances::free_balance(6), 10);
-		assert_eq!(Balances::free_balance(7), 5);
-	});
-}
-
-#[test]
-fn multisig_2_of_3_cannot_reissue_same_call() {
-	new_test_ext().execute_with(|| {
-		let multi = Utility::multi_account_id(&[1, 2, 3][..], 2);
-		assert_ok!(Balances::transfer(Origin::signed(1), multi, 5));
-		assert_ok!(Balances::transfer(Origin::signed(2), multi, 5));
-		assert_ok!(Balances::transfer(Origin::signed(3), multi, 5));
-
-		let call = Box::new(Call::Balances(BalancesCall::transfer(6, 10)));
-		assert_ok!(Utility::as_multi(Origin::signed(1), 2, vec![2, 3], None, call.clone()));
-		assert_ok!(Utility::as_multi(Origin::signed(2), 2, vec![1, 3], Some(now()), call.clone()));
-		assert_eq!(Balances::free_balance(multi), 5);
-
-		assert_ok!(Utility::as_multi(Origin::signed(1), 2, vec![2, 3], None, call.clone()));
-		assert_ok!(Utility::as_multi(Origin::signed(3), 2, vec![1, 2], Some(now()), call.clone()));
-
-		let err = DispatchError::from(BalancesError::<Test, _>::InsufficientBalance).stripped();
-		expect_event(RawEvent::MultisigExecuted(3, now(), multi, call.using_encoded(blake2_256), Err(err)));
-	});
-}
-
-#[test]
-fn zero_threshold_fails() {
-	new_test_ext().execute_with(|| {
-		let call = Box::new(Call::Balances(BalancesCall::transfer(6, 15)));
-		assert_noop!(
-			Utility::as_multi(Origin::signed(1), 0, vec![2], None, call),
-			Error::<Test>::ZeroThreshold,
-		);
-	});
-}
-
-#[test]
-fn too_many_signatories_fails() {
-	new_test_ext().execute_with(|| {
-		let call = Box::new(Call::Balances(BalancesCall::transfer(6, 15)));
-		assert_noop!(
-			Utility::as_multi(Origin::signed(1), 2, vec![2, 3, 4], None, call.clone()),
-			Error::<Test>::TooManySignatories,
-		);
-	});
-}
-
-#[test]
-fn duplicate_approvals_are_ignored() {
-	new_test_ext().execute_with(|| {
-		let call = Box::new(Call::Balances(BalancesCall::transfer(6, 15)));
-		let hash = call.using_encoded(blake2_256);
-		assert_ok!(Utility::approve_as_multi(Origin::signed(1), 2, vec![2, 3], None, hash.clone()));
-		assert_noop!(
-			Utility::approve_as_multi(Origin::signed(1), 2, vec![2, 3], Some(now()), hash.clone()),
-			Error::<Test>::AlreadyApproved,
-		);
-		assert_ok!(Utility::approve_as_multi(Origin::signed(2), 2, vec![1, 3], Some(now()), hash.clone()));
-		assert_noop!(
-			Utility::approve_as_multi(Origin::signed(3), 2, vec![1, 2], Some(now()), hash.clone()),
-			Error::<Test>::NoApprovalsNeeded,
-		);
-	});
-}
-
-#[test]
-fn multisig_1_of_3_works() {
-	new_test_ext().execute_with(|| {
-		let multi = Utility::multi_account_id(&[1, 2, 3][..], 1);
-		assert_ok!(Balances::transfer(Origin::signed(1), multi, 5));
-		assert_ok!(Balances::transfer(Origin::signed(2), multi, 5));
-		assert_ok!(Balances::transfer(Origin::signed(3), multi, 5));
-
-		let call = Box::new(Call::Balances(BalancesCall::transfer(6, 15)));
-		let hash = call.using_encoded(blake2_256);
-		assert_noop!(
-			Utility::approve_as_multi(Origin::signed(1), 1, vec![2, 3], None, hash.clone()),
-			Error::<Test>::NoApprovalsNeeded,
-		);
-		assert_noop!(
-			Utility::as_multi(Origin::signed(4), 1, vec![2, 3], None, call.clone()),
-			BalancesError::<Test, _>::InsufficientBalance,
-		);
-		assert_ok!(Utility::as_multi(Origin::signed(1), 1, vec![2, 3], None, call));
-
-		assert_eq!(Balances::free_balance(6), 15);
-	});
-}
-
-#[test]
-fn multisig_filters() {
-	new_test_ext().execute_with(|| {
-		let call = Box::new(Call::System(frame_system::Call::remark(vec![])));
-		assert_noop!(
-			Utility::as_multi(Origin::signed(1), 1, vec![], None, call.clone()),
-			Error::<Test>::Uncallable,
-		);
-	});
 }
 
 #[test]
